Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 # include <stdio.h>
 int main() {
-    int num1,num2,num3;
+    float num1,num2,num3;
     float avg;
-    scanf("%d %d %d",&num1,&num2,&num3);
+    scanf("%f %f %f",&num1,&num2,&num3);
     avg=(num1+num2+num3)/3;
     printf("Average: %.2f",avg);
     return 0;
